+++ src/1_HOY/src/boot_sect.bin
@@ -1,0 +1,7 @@
+loop: 
+
+    jmp loop
+
+times 510-($-$$) db 0
+
+dw 0xaa55
